requirements.txt
Django==1.9.6
django-cors-middleware==1.2.0
djangorestframework==3.3.3
httplib2==0.9.2
PySocks==1.5.7
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
twilio==5.4.0

